Add a script to generate the Markdown
The major addition is that we now have a `posts` and `pages`
top level sub-directories that are automatically generated
from Markdown files, using a simple shell script. This matches
what other micro-blogs do.

The Docker image was also updated to run the expected steps.

I hope this will simplify the logic down the road.